+++ honey.py
@@ -1,0 +1,4 @@
+from honeyssh.server import start_server
+
+if __name__ == "__main__":
+    start_server()
