+++ honeyssh/__init__.py
@@ -1,0 +1,3 @@
+from .server import start_server, main
+
+__all__ = ["start_server", "main"]
